Add files via upload
This is a Supervised Machine Learning exercise on studying the voice samples submitted for a Singing Show. 
The first task for Data Analyst/Scientist is to classify the voices (in the .WAV format) into a male/female category.

This is done by first classifying the voice parameters on the basis of frequency and then analyzing the data.
And here, we make us of Supervised M/c Learning algo and check the accuracy of our model.